Added a plone41 migration profile.
This is applied during an upgrade step of Products.minaraad.

Here we can put import steps that should be run once for the
migration.  Currently just the skins.xml.

We add the new mina_styles layer only to the Minaraad Theme there.

Increased the buildout version to 5.0 and the Products.minaraad
profile to 5000.  That leaves room for profile updates in maintenance
(3.x) and current trunk (4.x).

svn path=/minaraad/branches/plone4/; revision=30752