Move i18n-overrides part below the debugclient part.
Otherwise buildout will quit when installing i18n-overrides,
because parts/debugclients does not exist yet.
Updated this recipe to 1.2, which has better error handling so you actually notice this error.
I was wondering why the omelette part was not installed, and this was because the recipe quit with a sys.exit(1),
which should have been 'raise zc.buildout.UserError'.
Anyway, we should not be using this recipe anymore, but that is for another commit.

svn path=/minaraad/trunk/; revision=31484