Refuse to send an e-mail for newsletters, etcetera, twice.
Or rather, do not allow this when this e-mail has already been sent in
the last thirty minutes.  This will raise an exception and show a
traceback.  We also show a warning on the email_out page with the date
and time when this item was sent.  Sending a test e-mail is always
possible.

svn path=/minaraad/trunk/; revision=31697