Revert 'Try to fix functional test isolation problem.'
I thought it helped, but actually it did not.
Still fine on Mac.  Has been going wrong on Linux for years as far as I can see.

It may be better to switch to plone.app.testing,
which has been setup with layers and test isolation in mind.
But we can wait with that till we update to Plone 4.3.

svn path=/minaraad/trunk/; revision=32112