Added RedirectPlone helper view.
When using this, you can point to for example /@@newsletter in the
diazo theme and still end up in /plonesite/@@newsletter when used
locally.  Just register a browser page for the Zope Application root
with the name you need.

This is indeed done for @@newsletter now.

This refs issue #MINA-78.