Moved content/themes.py to ThemeMixin.py.
Renamed the main items to old_theme_schema and OldThemeMixin.
There were too many ThemeMixin classes and two themes.py files.

This reshuffle makes it easier for me to see which is old and which is new.
And it keeps all theme related field names in one spot,
which is good because I will be adding new ones for a new ThemeReferenceMixin.