Show/hide un/subscribe button and form with javascript.
By default, show the subscribe button and form.
The form is only filled in when you are a member, unless it is a POST request, both of which are not cached.
The template has a list of hashes of attendees in a script tag.
In javascript we read this and the cookie.
We show/hide stuff based on this.
And we fill the form using the cookie, if nothing is filled in yet.