Use much newer zc.buildout, setuptools and bootstrap.py versions.
You need to run python2.7 bootstrap.py.

Removed the buildout.dumppickedversions extension.  On the old
zc.buildout it failed to see the Jinja pin, and wanted it lowercase
instead (which is basically the 'fault' of mr.developer).  Good enough
reason to switch to newer buildout that does not have these problems.