Added /++theme++minaraad/static in front of font paths.
This is needed since we no longer load static/main.css in index.html
but in portal_javascript.  That means it gets combined with other css
and gets a different path.

Downside: index.html on its own no longer shows the fonts,
because it has no way to find them with the new path.