Replaced another strange character in HISTORY.txt.
zest.releaser can't make heads or tails from this during prerelease
and turns into into a different, also wrong, character.  So I'll fix
it here into an 'e' with an accent.  But then prerelease still
stumbles over it, turning it into the original ugly character.  It
*should* be able to handle it, if the file is in a sane format, like
utf-8, which maybe it is not?  Solution: use a standard 'e' without
accent.