Take over speedy project_workflow migration changes in real profile.
Products.minaraad has a migration that changes the project_workflow,
so we want to have this same file in minaraad.projects,
for clarity, for correct behavior in tests, and for i18ndude.

The migration one was missing i18n markup, which is okay for the
migration, but we need the markup when we run i18ndude.  So I copied
the updated file from minaraad.projects back to the migration.